implemented normal d3-plugin workflow
