+++ index.js
@@ -1,1 +1,1 @@
-export { default as orgChart } from "./src/d3-org-chart";
+export { default as OrgChart } from "./src/d3-org-chart";
