+++ package.json
@@ -22,7 +22,7 @@
     "url": "https://github.com/bumbeishvili/d3-organization-chart.git"
   },
   "scripts": {
-    "pretest": "rm -rf build && mkdir build && rollup -f umd -n d3 -o build/d3-org-chart.js -- index.js",
+    "pretest": "rm -rf build && mkdir build && rollup -g d3-selection:d3,d3-array:d3,d3-hierarchy:d3,d3-zoom:d3  -f umd -n d3 -o build/d3-org-chart.js -- index.js",
     "test": "tape 'test/**/*-test.js'",
     "prepublish": "npm run test && uglifyjs build/d3-org-chart.js -c -m -o build/d3-org-chart.min.js",
     "postpublish": "zip -j build/d3-org-chart.zip -- LICENSE README.md build/d3-org-chart.js build/d3-org-chart.min.js"
@@ -30,6 +30,12 @@
   "devDependencies": {
     "rollup": "0.27",
     "tape": "4",
-    "uglify-js": "2"
+    "uglify-js": "3"
+  },
+  "dependencies": {
+    "d3-selection": "3",
+    "d3-array": "3",
+    "d3-hierarchy": "3",
+    "d3-zoom": "3"
   }
 }
